Extract 15993 repro
