+++ e2e/test/scenarios/dashboard-cards/click-behavior.cy.spec.js
@@ -1,86 +1,24 @@
+import { WRITABLE_DB_ID } from "e2e/support/cypress_data";
+import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
+import { ORDERS_QUESTION_ID } from "e2e/support/cypress_sample_instance_data";
 import {
   addOrUpdateDashboardCard,
-  restore,
-  resetTestTable,
-  resyncDatabase,
-  visitDashboard,
   editDashboard,
+  resetTestTable,
+  restore,
+  resyncDatabase,
   showDashboardCardActions,
   sidebar,
+  visitDashboard,
 } from "e2e/support/helpers";
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import { WRITABLE_DB_ID } from "e2e/support/cypress_data";
-import { ORDERS_QUESTION_ID } from "e2e/support/cypress_sample_instance_data";
 
-const { ORDERS, ORDERS_ID, PRODUCTS, PRODUCTS_ID, REVIEWS, REVIEWS_ID } =
-  SAMPLE_DATABASE;
+const { PRODUCTS, PRODUCTS_ID, REVIEWS, REVIEWS_ID } = SAMPLE_DATABASE;
 
 describe("scenarios > dashboard > dashboard cards > click behavior", () => {
   beforeEach(() => {
     restore();
     cy.signInAsAdmin();
     cy.intercept("POST", "/api/dataset").as("dataset");
-  });
-
-  it("should show filters defined on a question with filter pass-thru (metabase#15993)", () => {
-    cy.createQuestion({
-      name: "15993",
-      query: {
-        "source-table": ORDERS_ID,
-      },
-    }).then(({ body: { id: question1Id } }) => {
-      cy.createNativeQuestion({ native: { query: "select 0" } }).then(
-        ({ body: { id: nativeId } }) => {
-          cy.createDashboard().then(({ body: { id: dashboardId } }) => {
-            // Add native question to the dashboard
-            addOrUpdateDashboardCard({
-              dashboard_id: dashboardId,
-              card_id: nativeId,
-              card: {
-                // Add click behavior to the dashboard card and point it to the question 1
-                visualization_settings: getVisualizationSettings(question1Id),
-              },
-            });
-            visitDashboard(dashboardId);
-          });
-        },
-      );
-    });
-
-    // Drill-through
-    cy.findAllByTestId("cell-data").get(".link").contains("0").realClick();
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.contains("117.03").should("not.exist"); // Total for the order in which quantity wasn't 0
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Quantity is equal to 0");
-
-    const getVisualizationSettings = targetId => ({
-      column_settings: {
-        '["name","0"]': {
-          click_behavior: {
-            targetId,
-            parameterMapping: {
-              [`["dimension",["field",${ORDERS.QUANTITY},null]]`]: {
-                source: {
-                  type: "column",
-                  id: "0",
-                  name: "0",
-                },
-                target: {
-                  type: "dimension",
-                  id: [`["dimension",["field",${ORDERS.QUANTITY},null]]`],
-                  dimension: ["dimension", ["field", ORDERS.QUANTITY, null]],
-                },
-                id: [`["dimension",["field",${ORDERS.QUANTITY},null]]`],
-              },
-            },
-            linkType: "question",
-            type: "link",
-          },
-        },
-      },
-    });
   });
 
   it("should not change the visualization type in a targetted question with mapped filter (metabase#16334)", () => {
